pyproject.toml
[project]
name = "talpred-ai"
version = "0.1.0"
description = "Stock prediction AI platform - Market Intelligence Predictor"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.4.0",
    "shap>=0.45.0",
    "yfinance>=0.2.36",
    "fredapi>=0.5.0",
    "requests>=2.31.0",
    "transformers>=4.38.0",
    "torch>=2.2.0",
    "prefect>=2.14.0",
    "mlflow>=2.10.0",
    "pandas-market-calendars>=4.3.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
